requirements.txt
confluent-kafka
fhir.resources>=7.0.0
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
requests
pandas
numpy
PyJWT
passlib
bcrypt
python-multipart
python-jose[cryptography]
